services/executor-py/requirements.txt
MetaTrader5==5.0.45
redis==5.2.1
requests==2.32.3
pydantic==2.10.6
python-dotenv==1.0.1
